+++ src/application/use-cases/station/CreateStationUseCase.ts
@@ -1,6 +1,6 @@
 import { create } from "domain";
-import { IStationRepository } from "../../../domain/models/entities/Station";
 import CreateStationDTO from "../../../web/dtos/station/CreateStationDTO";
+import { IStationRepository } from "../../../domain/interfaces/repositories/IStationRepository";
 
 export class CreateStationUseCase {
   constructor(private stationRepository: IStationRepository) {}
