+++ src/application/use-cases/station/UpdateStationUseCase.ts
@@ -1,4 +1,5 @@
-import { IStationRepository, Station } from "../../../domain/models/entities/Station";
+import { IStationRepository } from "../../../domain/interfaces/repositories/IStationRepository";
+import { Station } from "../../../domain/models/entities/Station";
 import UpdateStationDTO from "../../../web/dtos/station/UpdateStationDTO";
 
 export default class UpdateStationUseCase {
